# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "speedFlags API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/countries": {
      "get": {
        "summary": "Countries",
        "operationId": "countries_api_v1_countries_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryList"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games": {
      "post": {
        "summary": "Create",
        "operationId": "create_api_v1_games_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Settings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Gone",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}/start": {
      "post": {
        "summary": "Start",
        "operationId": "start_api_v1_games__game_id__start_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/games/{game_id}/answers": {
      "post": {
        "summary": "Answer",
        "operationId": "answer_api_v1_games__game_id__answers_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/games/{game_id}/sync": {
      "post": {
        "summary": "Sync",
        "operationId": "sync_api_v1_games__game_id__sync_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/games/{game_id}/finish": {
      "post": {
        "summary": "Finish",
        "operationId": "finish_api_v1_games__game_id__finish_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnswerRequest": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 65536,
            "minLength": 40,
            "title": "Token"
          },
          "question_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Question Id"
          },
          "sequence": {
            "type": "integer",
            "maximum": 250.0,
            "minimum": 0.0,
            "title": "Sequence"
          },
          "submission_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9-]{8,80}$",
            "title": "Submission Id"
          },
          "answer": {
            "type": "string",
            "maxLength": 100,
            "title": "Answer",
            "default": ""
          },
          "skip": {
            "type": "boolean",
            "title": "Skip",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "question_id",
          "sequence",
          "submission_id"
        ],
        "title": "AnswerRequest"
      },
      "Attempt": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "asset_url": {
            "type": "string",
            "title": "Asset Url"
          },
          "country_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Country Ids"
          },
          "accepted_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accepted Names"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "result": {
            "type": "string",
            "enum": [
              "correct",
              "incorrect",
              "skipped"
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "asset_url",
          "country_ids",
          "accepted_names",
          "answer",
          "result"
        ],
        "title": "Attempt"
      },
      "ContextRequest": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 65536,
            "minLength": 40,
            "title": "Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token"
        ],
        "title": "ContextRequest"
      },
      "Country": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "aliases"
        ],
        "title": "Country"
      },
      "CountryList": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "countries": {
            "items": {
              "$ref": "#/components/schemas/Country"
            },
            "type": "array",
            "title": "Countries"
          }
        },
        "type": "object",
        "required": [
          "version",
          "countries"
        ],
        "title": "CountryList"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "GameResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "playing",
              "finished"
            ],
            "title": "Status"
          },
          "settings": {
            "$ref": "#/components/schemas/Settings"
          },
          "server_time": {
            "type": "number",
            "title": "Server Time"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Question"
              },
              {
                "type": "null"
              }
            ]
          },
          "warmup_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warmup Answer"
          },
          "last_attempt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Attempt"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Attempt"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "time",
                  "deck",
                  "ended",
                  "session"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason"
          },
          "eligible_best": {
            "type": "boolean",
            "title": "Eligible Best",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "token",
          "dataset_version",
          "status",
          "settings",
          "server_time",
          "deadline",
          "score",
          "attempts",
          "skipped",
          "question"
        ],
        "title": "GameResponse"
      },
      "Question": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "asset_url": {
            "type": "string",
            "title": "Asset Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sequence",
          "asset_url"
        ],
        "title": "Question"
      },
      "Settings": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "timed",
              "practice"
            ],
            "title": "Mode",
            "default": "timed"
          },
          "duration": {
            "type": "integer",
            "enum": [
              30,
              45,
              60,
              120
            ],
            "title": "Duration",
            "default": 30
          },
          "bonus": {
            "type": "integer",
            "enum": [
              0,
              5
            ],
            "title": "Bonus",
            "default": 0
          },
          "scope": {
            "type": "string",
            "enum": [
              "starter",
              "all"
            ],
            "title": "Scope",
            "default": "starter"
          },
          "country_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 250,
            "title": "Country Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Settings"
      },
      "StartRequest": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 65536,
            "minLength": 40,
            "title": "Token"
          },
          "answer": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Answer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "answer"
        ],
        "title": "StartRequest"
      }
    }
  }
}